refactor(ngcc): break entry-point finding into composable units
This commit changes the simple `TracingEntryPointFinder` hierarchy
so that the two relevant parts (`InitialEntryPoints` and
`EntryPointsWithDeps`) are now composable.
This allows implementation  of a more flexible combination of
`EntryPointFinder` strategies.
