+++ packages/compiler-cli/ngcc/src/entry_point_finder/initial_entry_points/single_initial_entry_points.ts
@@ -1,0 +1,17 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+import {AbsoluteFsPath} from '../../../../src/ngtsc/file_system';
+import {InitialEntryPoints} from './interface';
+
+export class SingleInitialEntryPoints implements InitialEntryPoints {
+  constructor(private targetPath: AbsoluteFsPath) {}
+
+  getInitialEntryPointPaths(): AbsoluteFsPath[] {
+    return [this.targetPath];
+  }
+}
