+++ packages/compiler-cli/ngcc/src/entry_point_finder/tracing_entry_point_finder.ts
@@ -5,16 +5,13 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {AbsoluteFsPath, FileSystem} from '../../../src/ngtsc/file_system';
-import {Logger} from '../../../src/ngtsc/logging';
-
+import {AbsoluteFsPath} from '../../../src/ngtsc/file_system';
 import {EntryPointWithDependencies} from '../dependencies/dependency_host';
 import {DependencyResolver, SortedEntryPointsInfo} from '../dependencies/dependency_resolver';
-import {NgccConfiguration} from '../packages/configuration';
-import {PathMappings} from '../path_mappings';
 
+import {EntryPointsWithDeps} from './entry_points_with_deps/interface';
+import {InitialEntryPoints} from './initial_entry_points/interface';
 import {EntryPointFinder} from './interface';
-import {getBasePaths} from './utils';
 
 /**
  * An EntryPointFinder that starts from a set of initial files and only returns entry-points that
@@ -33,22 +30,19 @@ import {getBasePaths} from './utils';
  *   where the `tsconfig.json` to be used to do the build is known.
  */
 export abstract class TracingEntryPointFinder implements EntryPointFinder {
-  private basePaths: AbsoluteFsPath[]|null = null;
-
   constructor(
-      protected fs: FileSystem, protected config: NgccConfiguration, protected logger: Logger,
-      protected resolver: DependencyResolver, protected basePath: AbsoluteFsPath,
-      protected pathMappings: PathMappings|undefined) {}
+      protected resolver: DependencyResolver, protected initialEntryPoints: InitialEntryPoints,
+      protected entryPointsWithDeps: EntryPointsWithDeps) {}
 
   /**
    * Search for Angular package entry-points.
    */
   findEntryPoints(): SortedEntryPointsInfo {
     const unsortedEntryPoints = new Map<AbsoluteFsPath, EntryPointWithDependencies>();
-    const unprocessedPaths = this.getInitialEntryPointPaths();
+    const unprocessedPaths = this.initialEntryPoints.getInitialEntryPointPaths();
     while (unprocessedPaths.length > 0) {
       const path = unprocessedPaths.shift()!;
-      const entryPointWithDeps = this.getEntryPointWithDeps(path);
+      const entryPointWithDeps = this.entryPointsWithDeps.getEntryPointWithDeps(path);
       if (entryPointWithDeps === null) {
         continue;
       }
@@ -60,39 +54,5 @@ export abstract class TracingEntryPointFinder implements EntryPointFinder {
       });
     }
     return this.resolver.sortEntryPointsByDependency(Array.from(unsortedEntryPoints.values()));
-  }
-
-
-  /**
-   * Return an array of entry-point paths from which to start the trace.
-   */
-  protected abstract getInitialEntryPointPaths(): AbsoluteFsPath[];
-
-  /**
-   * For the given `entryPointPath`, compute, or retrieve, the entry-point information, including
-   * paths to other entry-points that this entry-point depends upon.
-   *
-   * @param entryPointPath the path to the entry-point whose information and dependencies are to be
-   *     retrieved or computed.
-   *
-   * @returns the entry-point and its dependencies or `null` if the entry-point is not compiled by
-   *     Angular or cannot be determined.
-   */
-  protected abstract getEntryPointWithDeps(entryPointPath: AbsoluteFsPath):
-      EntryPointWithDependencies|null;
-
-
-  /**
-   * Parse the path-mappings to compute the base-paths that need to be considered when finding
-   * entry-points.
-   *
-   * This processing can be time-consuming if the path-mappings are complex or extensive.
-   * So the result is cached locally once computed.
-   */
-  protected getBasePaths() {
-    if (this.basePaths === null) {
-      this.basePaths = getBasePaths(this.logger, this.basePath, this.pathMappings);
-    }
-    return this.basePaths;
   }
 }
