+++ packages/compiler-cli/ngcc/test/entry_point_finder/program_based_entry_point_finder_spec.ts
@@ -10,11 +10,11 @@ import {runInEachFileSystem, TestFile} from '../../../src/ngtsc/file_system/test
 import {MockLogger} from '../../../src/ngtsc/logging/testing';
 import {readConfiguration} from '../../../src/perform_compile';
 import {loadTestFiles} from '../../../test/helpers';
-
 import {DependencyResolver} from '../../src/dependencies/dependency_resolver';
 import {DtsDependencyHost} from '../../src/dependencies/dts_dependency_host';
 import {EsmDependencyHost} from '../../src/dependencies/esm_dependency_host';
 import {ModuleResolver} from '../../src/dependencies/module_resolver';
+import {BasePaths} from '../../src/entry_point_finder/base_paths';
 import {EntryPointCollector} from '../../src/entry_point_finder/entry_point_collector';
 import {ProgramBasedEntryPointFinder} from '../../src/entry_point_finder/program_based_entry_point_finder';
 import {NgccConfiguration} from '../../src/packages/configuration';
@@ -72,7 +72,8 @@ runInEachFileSystem(() => {
         const collector = new EntryPointCollector(fs, config, logger, resolver);
         const manifest = new EntryPointManifest(fs, config, logger);
         return new ProgramBasedEntryPointFinder(
-            fs, config, logger, resolver, collector, manifest, basePath, tsConfig, projectPath);
+            fs, logger, resolver, collector, manifest, basePath, new BasePaths(logger), tsConfig,
+            projectPath);
       }
 
       function createProgram(projectPath: AbsoluteFsPath): TestFile[] {
